fix(flutter): repair the harvested ratings/personals batch
The harvested commit did not analyze or test clean. Fixes, in order of
severity:

Data loss, both from drift writing only the columns a companion carries:

- PersonalsRepository.update used `toCompanion(true)` (nullToAbsent), so
  clearing a note's description dropped the column from the statement and
  left the old text in place. Room's `@Update` in PersonalDao writes the
  whole row; `toCompanion(false)` matches it. SessionNotifier.updateProfile
  had the same bug — blanking a phone number or email never persisted.
- NotificationsRepository never reset `isRead` on upsert, so a changed
  resource or storage count stayed marked read and the badge went stale.
  NotificationsRepositoryImpl.kt sets `isRead = false` whenever the count
  changes; the companion now passes it explicitly.

Tests that were asserting against nothing:

- wrapScreen now redirects appDatabaseProvider to an in-memory database
  ahead of the caller's overrides. ProfileScreen's unread badge and
  CourseDetailScreen's rating summary were reaching AppDatabase.open(),
  whose path_provider lookup has no platform channel under `flutter test`;
  both screens read the result through `.valueOrNull ?? <default>`, so the
  failure was swallowed and the tests passed. Both now stub the provider
  they actually read.

Build and analyze:

- dictionary_mapper.dart imported tables.dart, where the generated
  companions do not live (they are part of app_database.dart).
- profile_screen_test.dart declared imports below its declarations.
- Dropped an unnecessary cast, a deprecated `onReorder`, and an unused
  import; formatted the 17 files the harvest left unformatted.

Test expectations corrected rather than the code: life_screen_test expected
dashboard-visibility prose the ARB does not carry, dictionary_screen_test
tapped a capitalized 'Search' where the string is verbatim-from-Kotlin
lowercase, and its synonym assertion could not match a Text.rich span
without findRichText.

Localization: 35 of the batch's new keys had an unambiguous values-es
counterpart and are now translated — carried over only where the Kotlin
string name and its English text both agree. The remaining 102 are new
phrasings with no Kotlin original and stay English-only for Crowdin.

226 tests pass, `flutter analyze` is clean, and the debug APK builds.

+++ flutter/lib/repository/ratings_repository.dart
@@ -73,7 +73,7 @@ class RatingsRepository {
         title: Value(title),
         userId: userId,
         isUpdated: const Value(true),
-        rate: rate.clamp(1, 5) as int,
+        rate: rate.clamp(1, 5),
         item: itemId,
         comment: Value(_nullable(comment)),
         parentCode: Value(parentCode),
